+++ src/main/simplejpa/swing/glazed/GlazedColumn.groovy
@@ -16,7 +16,11 @@
 
 package simplejpa.swing.glazed
 
+import groovy.beans.Bindable
+import javax.swing.event.SwingPropertyChangeSupport
 import javax.swing.table.TableColumn
+import java.beans.PropertyChangeListener
+
 
 class GlazedColumn extends TableColumn {
 
@@ -29,5 +33,54 @@ class GlazedColumn extends TableColumn {
     String property
     Class columnClass
     Comparator comparator
+    @Bindable Boolean visible = true
+
+    //
+    // @Bindable can't generate this because parent already has these methods,
+    // but we can't use them because some of them are declared private by parents.
+    // So just copied from parent to this here.
+    //
+    private SwingPropertyChangeSupport changeSupport
+
+    public synchronized void addPropertyChangeListener(PropertyChangeListener listener) {
+        super.addPropertyChangeListener(listener)
+        if (changeSupport == null) changeSupport = new SwingPropertyChangeSupport(this)
+        changeSupport.addPropertyChangeListener(listener)
+    }
+
+    public synchronized void removePropertyChangeListener(PropertyChangeListener listener) {
+        super.removePropertyChangeListener(listener)
+        if (changeSupport != null) {
+            changeSupport.removePropertyChangeListener(listener);
+        }
+    }
+
+    public synchronized PropertyChangeListener[] getPropertyChangeListeners() {
+        if (changeSupport == null) return new PropertyChangeListener[0]
+        return changeSupport.getPropertyChangeListeners()
+    }
+
+
+    private void firePropertyChange(String propertyName, Object oldValue, Object newValue) {
+        if (changeSupport != null) {
+            changeSupport.firePropertyChange(propertyName, oldValue, newValue);
+        }
+    }
+
+    private void firePropertyChange(String propertyName, int oldValue, int newValue) {
+        if (oldValue != newValue) {
+            firePropertyChange(propertyName, Integer.valueOf(oldValue), Integer.valueOf(newValue));
+        }
+    }
+
+    private void firePropertyChange(String propertyName, boolean oldValue, boolean newValue) {
+        if (oldValue != newValue) {
+            firePropertyChange(propertyName, Boolean.valueOf(oldValue), Boolean.valueOf(newValue));
+        }
+    }
+
+    //
+    // End of property change listener
+    //
 
 }